Add files via upload
A clock is to be designed. The clock can be started and stopped (EdgeDetect) by pressing a button (startStop). The clock should count in milliseconds. Minutes (HEX5), seconds (HEX4, HEX3) and milliseconds (HEX2..0) are displayed. The decimal point lights up between minutes and seconds and between seconds and milliseconds. The minutes run from 0 to 9, the seconds run from 0 to 59, and the milliseconds run from 0 to 999. #VHDL